Crate config_file_handler

Config File Handler

Create, read and write JSON-encoded config files.

Structs

FileHandler
Struct for reading and writing config files.

ScopedUserAppDirRemover
RAII object which removes the user_app_dir() when an instance is dropped.

Enums

Error
Error types.

Functions

current_bin_dir
The full path to the directory containing the currently-running binary. See also an example config file flowchart.

exe_file_stem
The file name of the currently-running binary without any suffix or extension. For example, if the binary is "C:\Abc.exe" this function will return Ok("Abc").

system_cache_dir
The full path to a system cache directory available for all users. See also an example config file flowchart.

user_app_dir
The full path to an application support directory for the current user. See also an example config file flowchart.